perf: stats table rendering
